+++ lib/presentation/pages/widgets/marker_infowindow.dart
@@ -24,7 +24,7 @@ class CustomMarkerPopup extends StatefulWidget {
   final String stationName;
   final String address;
   final List<dynamic> imageUrl;
-  final double costOfFullCharge;
+  final int costOfFullCharge;
   final num startTime;
   final num endTime;
   final int timeslot;
