+++ lib/presentation/register/vForm.dart
@@ -22,7 +22,7 @@ class VForm extends StatefulWidget {
   final String stationName;
   final String address;
   final List<dynamic> imageUrl;
-  final double costOfFullCharge;
+  final int costOfFullCharge;
   final num startTime;
   final num endTime;
   final int timeslot;
